+++ core/api/typescript/src/lib/kurtosis_api_version_const.ts
@@ -1,2 +1,2 @@
 // DO NOT UPDATE THIS VALUE! It will get updated automatically during the release process
-export const KURTOSIS_API_VERSION: string = "1.30.0"
+export const KURTOSIS_API_VERSION: string = "1.31.0"