+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 1.31.0
 ### Changes
 * The API container launcher now has two methods, `LaunchWithCustomVersion` and `LaunchWithDefaultVersion`